Extract reusable methods into the EnumeratesValues trait
